llama_backend/pyproject.toml
[tool.poetry]
name = "llama-backend"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.12.8"
django = "^5.1.4"
djangorestframework = "^3.15.2"
django-filter = "^24.3"
markdown = "^3.7"
pillow = "^11.0.0"
celery = "^5.4.0"
django-celery-results = "^2.5.1"
django-redis = "^5.4.0"
django-celery-beat = "^2.7.0"
torch = "^2.5.1"
torchvision = "^0.20.1"
torchaudio = "^2.5.1"
transformers = "4.45.2"
llama-index = "^0.12.5"
python-dotenv = "^1.0.1"
llama-index-llms-huggingface = "^0.4.0"
accelerate = "^1.2.0"
mlx-lm = "^0.20.3"
llama-index-llms-ollama = "^0.5.0"
llama-index-embeddings-huggingface = "^0.4.0"
django-cors-headers = "^4.6.0"
django-background-tasks = "^1.2.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
